### 2018-01-02 #176 @erzk
PostcodesioR submission
It is an API wrapper for http://postcodes.io/ which is a free geolocation service for the UK data.

```
Package: PostcodesioR
Title: API wrapper around postcodes.io (free UK postcode lookup and geocoder)
Version: 0.1.1
Authors@R: person("Eryk", "Walczak", email = "eryk.j.walczak@gmail.com", role = c("aut", "cre"))
Description: Free UK geocoding using data from Office for National Statistics.
    It is using several functions to get information about post codes, outward
    codes, reverse geocoding, nearest post codes/outward codes, validation, or
    randomly generate a post code.
License: GPL-3 + file LICENSE
    R (>= 3.1),
    httr
Suggests: knitr,
    covr
```

https://github.com/erzk/PostcodesioR

Geospatial data, because the package deals with UK geospatial data.

Whoever needs information about UK postcodes, outcodes, or places. This package will be particularly useful for social scientists.

Not that I'm aware of. This package return a rich geospatial data related to postcodes (but not only).

NA

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.
