+++ R/zzz.R
@@ -1,2 +1,1 @@
-ua <- user_agent("http://github.com/erzk/PostcodesioR")
-ua
+ua <- httr::user_agent("http://github.com/erzk/PostcodesioR")
