+++ NAMESPACE
@@ -15,4 +15,5 @@ export(postcode_validation)
 export(random_place)
 export(random_postcode)
 export(reverse_geocoding)
+export(terminated_postcode)
 import(httr)
