+++ README.md
@@ -83,38 +83,38 @@ str(bulk_lookup_result$result[1])
 
     ## List of 1
     ##  $ :List of 2
-    ##   ..$ query : chr "EX165BL"
+    ##   ..$ query : chr "PR3 0SG"
     ##   ..$ result:List of 23
-    ##   .. ..$ postcode                  : chr "EX16 5BL"
+    ##   .. ..$ postcode                  : chr "PR3 0SG"
     ##   .. ..$ quality                   : int 1
-    ##   .. ..$ eastings                  : int 294480
-    ##   .. ..$ northings                 : int 112251
+    ##   .. ..$ eastings                  : int 351012
+    ##   .. ..$ northings                 : int 440302
     ##   .. ..$ country                   : chr "England"
-    ##   .. ..$ nhs_ha                    : chr "South West"
-    ##   .. ..$ longitude                 : num -3.5
-    ##   .. ..$ latitude                  : num 50.9
-    ##   .. ..$ european_electoral_region : chr "South West"
-    ##   .. ..$ primary_care_trust        : chr "Devon"
-    ##   .. ..$ region                    : chr "South West"
-    ##   .. ..$ lsoa                      : chr "Mid Devon 005C"
-    ##   .. ..$ msoa                      : chr "Mid Devon 005"
-    ##   .. ..$ incode                    : chr "5BL"
-    ##   .. ..$ outcode                   : chr "EX16"
-    ##   .. ..$ parliamentary_constituency: chr "Tiverton and Honiton"
-    ##   .. ..$ admin_district            : chr "Mid Devon"
-    ##   .. ..$ parish                    : chr "Tiverton"
-    ##   .. ..$ admin_county              : chr "Devon"
-    ##   .. ..$ admin_ward                : chr "Westexe"
-    ##   .. ..$ ccg                       : chr "NHS Northern, Eastern and Western Devon"
-    ##   .. ..$ nuts                      : chr "Devon CC"
+    ##   .. ..$ nhs_ha                    : chr "North West"
+    ##   .. ..$ longitude                 : num -2.75
+    ##   .. ..$ latitude                  : num 53.9
+    ##   .. ..$ european_electoral_region : chr "North West"
+    ##   .. ..$ primary_care_trust        : chr "North Lancashire Teaching"
+    ##   .. ..$ region                    : chr "North West"
+    ##   .. ..$ lsoa                      : chr "Wyre 006A"
+    ##   .. ..$ msoa                      : chr "Wyre 006"
+    ##   .. ..$ incode                    : chr "0SG"
+    ##   .. ..$ outcode                   : chr "PR3"
+    ##   .. ..$ parliamentary_constituency: chr "Wyre and Preston North"
+    ##   .. ..$ admin_district            : chr "Wyre"
+    ##   .. ..$ parish                    : chr "Myerscough and Bilsborrow"
+    ##   .. ..$ admin_county              : chr "Lancashire"
+    ##   .. ..$ admin_ward                : chr "Brock with Catterall"
+    ##   .. ..$ ccg                       : chr "NHS Morecambe Bay"
+    ##   .. ..$ nuts                      : chr "Lancaster and Wyre"
     ##   .. ..$ codes                     :List of 7
-    ##   .. .. ..$ admin_district            : chr "E07000042"
-    ##   .. .. ..$ admin_county              : chr "E10000008"
-    ##   .. .. ..$ admin_ward                : chr "E05003531"
-    ##   .. .. ..$ parish                    : chr "E04003055"
-    ##   .. .. ..$ parliamentary_constituency: chr "E14000996"
-    ##   .. .. ..$ ccg                       : chr "E38000129"
-    ##   .. .. ..$ nuts                      : chr "UKK43"
+    ##   .. .. ..$ admin_district            : chr "E07000128"
[…]
