+++ README.md
@@ -402,27 +402,100 @@ str(out_rev_geocode$result[1])
 
 #### Postcodes
 
-Generates a data frame with a random UK postcode and corresponding geographic information:
+Generates a list with a random UK postcode and corresponding geographic information:
 
 ``` r
 # without restrictions
 random_postcode()
 ```
 
-    ##   postcode quality eastings northings country          nhs_ha longitude
-    ## 1 IP14 3BW       1   600408    257797 England East of England 0.9299627
-    ##   latitude european_electoral_region primary_care_trust          region
-    ## 1   52.182                   Eastern            Suffolk East of England
-    ##               lsoa            msoa incode outcode
-    ## 1 Mid Suffolk 009A Mid Suffolk 009    3BW    IP14
-    ##   parliamentary_constituency admin_district  parish admin_county
-    ## 1            Bury St Edmunds    Mid Suffolk Buxhall      Suffolk
-    ##   admin_ward                          ccg    nuts admin_district
-    ## 1   Onehouse NHS Ipswich and East Suffolk Suffolk      E07000203
-    ##   admin_county admin_ward    parish parliamentary_constituency       ccg
-    ## 1    E10000029  E05007148 E04010548                  E14000613 E38000086
-    ##    nuts
-    ## 1 UKH14
+    ## $postcode
+    ## [1] "EN6 5LX"
+    ## 
+    ## $quality
+    ## [1] 1
+    ## 
+    ## $eastings
+    ## [1] 526533
+    ## 
+    ## $northings
+    ## [1] 201296
+    ## 
+    ## $country
+    ## [1] "England"
+    ## 
+    ## $nhs_ha
+    ## [1] "East of England"
+    ## 
+    ## $longitude
+    ## [1] -0.1706038
+    ## 
+    ## $latitude
+    ## [1] 51.69616
+    ## 
+    ## $european_electoral_region
+    ## [1] "Eastern"
+    ## 
+    ## $primary_care_trust
+    ## [1] "Hertfordshire"
+    ## 
+    ## $region
+    ## [1] "East of England"
+    ## 
+    ## $lsoa
+    ## [1] "Hertsmere 003B"
+    ## 
+    ## $msoa
+    ## [1] "Hertsmere 003"
+    ## 
+    ## $incode
+    ## [1] "5LX"
+    ## 
+    ## $outcode
+    ## [1] "EN6"
+    ## 
+    ## $parliamentary_constituency
+    ## [1] "Hertsmere"
+    ## 
+    ## $admin_district
+    ## [1] "Hertsmere"
+    ## 
+    ## $parish
+    ## [1] "Hertsmere, unparished area"
+    ## 
+    ## $admin_county
+    ## [1] "Hertfordshire"
+    ## 
+    ## $admin_ward
+    ## [1] "Potters Bar Oakmere"
+    ## 
+    ## $ccg
+    ## [1] "NHS Herts Valleys"
+    ## 
+    ## $nuts
+    ## [1] "Hertfordshire"
+    ## 
+    ## $codes
+    ## $codes$admin_district
+    ## [1] "E07000098"
+    ## 
+    ## $codes$admin_county
+    ## [1] "E10000015"
+    ## 
+    ## $codes$admin_ward
+    ## [1] "E05004758"
+    ## 
+    ## $codes$parish
+    ## [1] "E43000077"
+    ## 
+    ## $codes$parliamentary_constituency
+    ## [1] "E14000745"
+    ## 
+    ## $codes$ccg
+    ## [1] "E38000079"
+    ## 
+    ## $codes$nuts
+    ## [1] "UKH23"
 
 A randomly generated postcode can also belong to a particular outcode:
 
[…]
