### 2018-01-06 2d54169
df to list (avoids errors)
