+++ tests/testthat/test-random_postcode.R
@@ -6,8 +6,8 @@ test_that("random_postcode works as expected", {
   random_postcode_result1 <- random_postcode()
   random_postcode_result2 <- random_postcode(test_outcode)
 
-  expect_that(random_postcode_result1, is_a("data.frame"))
-  expect_that(random_postcode_result2, is_a("data.frame"))
+  expect_that(random_postcode_result1, is_a("list"))
+  expect_that(random_postcode_result2, is_a("list"))
 
   expect_equal(random_postcode_result2$outcode, test_outcode)
 })
