### 2018-01-06 3dfcb5e
extract_results added
