+++ R/postcode_query.R
@@ -23,6 +23,5 @@ postcode_query <- function(postcode, limit = 10) {
     stop("Please provide an integer lower than 100.")
   }
   r <- GET(paste0("https://api.postcodes.io/postcodes?q=", postcode, "&limit=", limit))
-  warn_for_status(r)
-  content(r)
+  extract_results(r)
 }
