### 2018-01-06 3b124d7
$results removed
