+++ tests/testthat/test-postcode_autocomplete.R
@@ -16,5 +16,5 @@ test_that("postcode_autocomplete works as expected", {
   expect_that(postcode_autocomplete_result1, is_a("list"))
   expect_that(postcode_autocomplete_result2, is_a("list"))
 
-  expect_equal(length(postcode_autocomplete_result2$result), correct_limit)
+  expect_equal(length(postcode_autocomplete_result2), correct_limit)
 })
