+++ R/postcode_autocomplete.R
@@ -23,6 +23,5 @@ postcode_autocomplete <- function(postcode, limit = 10) {
     stop("Please provide an integer lower than 100.")
   }
   r <- GET(paste0("https://api.postcodes.io/postcodes/", postcode, "/autocomplete", "?limit=", limit))
-  warn_for_status(r)
-  content(r)
+  extract_results(r)
 }
