+++ README.md
@@ -410,92 +410,92 @@ random_postcode()
 ```
 
     ## $postcode
-    ## [1] "L34 9EY"
+    ## [1] "SE16 3TP"
     ## 
     ## $quality
     ## [1] 1
     ## 
     ## $eastings
-    ## [1] 343716
+    ## [1] 534677
     ## 
     ## $northings
-    ## [1] 396737
+    ## [1] 178823
     ## 
     ## $country
     ## [1] "England"
     ## 
     ## $nhs_ha
-    ## [1] "North West"
+    ## [1] "London"
     ## 
     ## $longitude
-    ## [1] -2.849254
+    ## [1] -0.06145744
     ## 
     ## $latitude
-    ## [1] 53.46436
+    ## [1] 51.49232
     ## 
     ## $european_electoral_region
-    ## [1] "North West"
+    ## [1] "London"
     ## 
     ## $primary_care_trust
-    ## [1] "Knowsley"
+    ## [1] "Southwark"
     ## 
     ## $region
-    ## [1] "North West"
+    ## [1] "London"
     ## 
     ## $lsoa
-    ## [1] "Knowsley 005E"
+    ## [1] "Southwark 010D"
     ## 
     ## $msoa
-    ## [1] "Knowsley 005"
+    ## [1] "Southwark 010"
     ## 
     ## $incode
-    ## [1] "9EY"
+    ## [1] "3TP"
     ## 
     ## $outcode
-    ## [1] "L34"
+    ## [1] "SE16"
     ## 
     ## $parliamentary_constituency
-    ## [1] "Knowsley"
+    ## [1] "Bermondsey and Old Southwark"
     ## 
     ## $admin_district
-    ## [1] "Knowsley"
+    ## [1] "Southwark"
     ## 
     ## $parish
-    ## [1] "Knowsley"
+    ## [1] "Southwark, unparished area"
     ## 
     ## $admin_county
     ## NULL
     ## 
     ## $admin_ward
-    ## [1] "Prescot North"
+    ## [1] "South Bermondsey"
     ## 
     ## $ccg
-    ## [1] "NHS Knowsley"
+    ## [1] "NHS Southwark"
     ## 
     ## $nuts
-    ## [1] "East Merseyside"
+    ## [1] "Lewisham and Southwark"
     ## 
     ## $codes
     ## $codes$admin_district
-    ## [1] "E08000011"
+    ## [1] "E09000028"
     ## 
     ## $codes$admin_county
     ## [1] "E99999999"
     ## 
     ## $codes$admin_ward
-    ## [1] "E05010940"
+    ## [1] "E05000550"
     ## 
     ## $codes$parish
-    ## [1] "E04000018"
+    ## [1] "E43000218"
     ## 
     ## $codes$parliamentary_constituency
-    ## [1] "E14000775"
+    ## [1] "E14000553"
     ## 
     ## $codes$ccg
-    ## [1] "E38000091"
+    ## [1] "E38000171"
     ## 
     ## $codes$nuts
-    ## [1] "UKD71"
+    ## [1] "UKI44"
 
 A randomly generated postcode can also belong to a particular outcode:
 
@@ -505,16 +505,16 @@ random_postcode("N1")
 ```
 
     ## $postcode
-    ## [1] "N1 9DN"
+    ## [1] "N1 1QB"
     ## 
     ## $quality
     ## [1] 1
     ## 
     ## $eastings
-    ## [1] 530558
+    ## [1] 531543
     ## 
     ## $northings
-    ## [1] 183237
+    ## [1] 184395
     ## 
     ## $country
     ## [1] "England"
@@ -523,10 +523,10 @@ random_postcode("N1")
     ## [1] "London"
[…]
