+++ README.md
@@ -410,92 +410,92 @@ random_postcode()
 ```
 
     ## $postcode
-    ## [1] "SE16 3TP"
+    ## [1] "PO19 7HW"
     ## 
     ## $quality
     ## [1] 1
     ## 
     ## $eastings
-    ## [1] 534677
+    ## [1] 487352
     ## 
     ## $northings
-    ## [1] 178823
+    ## [1] 105472
     ## 
     ## $country
     ## [1] "England"
     ## 
     ## $nhs_ha
-    ## [1] "London"
+    ## [1] "South East Coast"
     ## 
     ## $longitude
-    ## [1] -0.06145744
+    ## [1] -0.7607198
     ## 
     ## $latitude
-    ## [1] 51.49232
+    ## [1] 50.84219
     ## 
     ## $european_electoral_region
-    ## [1] "London"
+    ## [1] "South East"
     ## 
     ## $primary_care_trust
-    ## [1] "Southwark"
+    ## [1] "West Sussex"
     ## 
     ## $region
-    ## [1] "London"
+    ## [1] "South East"
     ## 
     ## $lsoa
-    ## [1] "Southwark 010D"
+    ## [1] "Chichester 008A"
     ## 
     ## $msoa
-    ## [1] "Southwark 010"
+    ## [1] "Chichester 008"
     ## 
     ## $incode
-    ## [1] "3TP"
+    ## [1] "7HW"
     ## 
     ## $outcode
-    ## [1] "SE16"
+    ## [1] "PO19"
     ## 
     ## $parliamentary_constituency
-    ## [1] "Bermondsey and Old Southwark"
+    ## [1] "Chichester"
     ## 
     ## $admin_district
-    ## [1] "Southwark"
+    ## [1] "Chichester"
     ## 
     ## $parish
-    ## [1] "Southwark, unparished area"
+    ## [1] "Chichester"
     ## 
     ## $admin_county
-    ## NULL
+    ## [1] "West Sussex"
     ## 
     ## $admin_ward
-    ## [1] "South Bermondsey"
+    ## [1] "Chichester East"
     ## 
     ## $ccg
-    ## [1] "NHS Southwark"
+    ## [1] "NHS Coastal West Sussex"
     ## 
     ## $nuts
-    ## [1] "Lewisham and Southwark"
+    ## [1] "West Sussex (South West)"
     ## 
     ## $codes
     ## $codes$admin_district
-    ## [1] "E09000028"
+    ## [1] "E07000225"
     ## 
     ## $codes$admin_county
-    ## [1] "E99999999"
+    ## [1] "E10000032"
     ## 
     ## $codes$admin_ward
-    ## [1] "E05000550"
+    ## [1] "E05007605"
     ## 
     ## $codes$parish
-    ## [1] "E43000218"
+    ## [1] "E04009888"
     ## 
     ## $codes$parliamentary_constituency
-    ## [1] "E14000553"
+    ## [1] "E14000633"
     ## 
     ## $codes$ccg
-    ## [1] "E38000171"
+    ## [1] "E38000213"
     ## 
     ## $codes$nuts
-    ## [1] "UKI44"
+    ## [1] "UKJ27"
 
 A randomly generated postcode can also belong to a particular outcode:
 
@@ -505,16 +505,16 @@ random_postcode("N1")
 ```
 
     ## $postcode
-    ## [1] "N1 1QB"
+    ## [1] "N1 2XL"
     ## 
     ## $quality
     ## [1] 1
     ## 
     ## $eastings
-    ## [1] 531543
+    ## [1] 531734
     ## 
     ## $northings
-    ## [1] 184395
+    ## [1] 184490
     ## 
     ## $country
[…]
