### 2018-01-10 #176 @karthik
### Editor checks:

---

#### Editor comments

Thanks for this submission to the geospatial area. While I seek reviewers, please look over some checks from `goodpractice` below. Not all will be relevant (e.g. test coverage is already quite high) but there are other issues worth fixing before the reviewers begin.

---

Reviewers: Currently seeking reviewers
Due date: TBD

```
── GP PostcodesioR ───────────────────────────────────────────────
It is good practice to

    R/outcode_reverse_geocoding.R:27:NA
    R/place_lookup.R:18:NA
    R/place_query.R:20:NA

  ✖ not use "Depends" in DESCRIPTION, as it can cause name
    clashes, and poor interaction with other packages. Use "Imports" instead.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easierto read your code for them if you use '<-'.

    R/postcode_lookup.R:26:41
    tests/testthat/test-place_query.R:6:17
    tests/testthat/test-postcode_autocomplete.R:7:17
    tests/testthat/test-postcode_query.R:8:17

    many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters

    R/outcode_reverse_geocoding.R:22:1
    R/place_query.R:25:1
    R/postcode_autocomplete.R:25:1
    R/postcode_query.R:25:1
    R/reverse_geocoding.R:22:1
    ... and 7 more lines

    vector, or a list, depending on the input data. Consider using vapply() instead.

    R/postcode_lookup.R:26:13
    R/random_place.R:19:21

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the specific functions you need.
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no
    redirection of stdout/stderr. Hmm ... looks like a package You may want to clean up by 'rm -rf /tmp/Rtmpbpsw7A/Rd2pdf59e56e855476'
───────────────────────────────────────────────```
