+++ R/place_query.R
@@ -22,6 +22,7 @@ place_query <- function(place, limit = 10) {
   if (limit > 100) {
     stop("Please provide an integer lower than 100.")
   }
-  r <- GET(paste0("https://api.postcodes.io/places?q=", place, "&limit=", limit))
+  r <- GET(paste0("https://api.postcodes.io/places?q=",
+                  place, "&limit=", limit))
   extract_results(r)
 }
