+++ R/postcode_autocomplete.R
@@ -22,6 +22,7 @@ postcode_autocomplete <- function(postcode, limit = 10) {
   if (limit > 100) {
     stop("Please provide an integer lower than 100.")
   }
-  r <- GET(paste0("https://api.postcodes.io/postcodes/", postcode, "/autocomplete", "?limit=", limit))
+  r <- GET(paste0("https://api.postcodes.io/postcodes/",
+                  postcode, "/autocomplete", "?limit=", limit))
   extract_results(r)
 }
