+++ tests/testthat/test-place_lookup.R
@@ -5,4 +5,6 @@ test_that("place_lookup works as expected", {
   place_lookup_result <- place_lookup("osgb4000000074544700")
 
   expect_that(place_lookup_result, is_a("list"))
+
+  expect_error(place_lookup(1))
 })
