+++ tests/testthat/test-postcode_autocomplete.R
@@ -4,10 +4,11 @@ test_that("postcode_autocomplete works as expected", {
   incorrect_postcode <- 1
   incorrect_limit1 <- 101
   incorrect_limit2 <- "11"
-  correct_limit = 11
+  correct_limit <- 11
 
   postcode_autocomplete_result1 <- postcode_autocomplete("E1")
-  postcode_autocomplete_result2 <- postcode_autocomplete("E1", limit = correct_limit)
+  postcode_autocomplete_result2 <-
+    postcode_autocomplete("E1", limit = correct_limit)
 
   expect_error(postcode_autocomplete(incorrect_postcode))
   expect_error(postcode_autocomplete("E1", limit = incorrect_limit1))
