+++ tests/testthat/test-reverse_geocoding.R
@@ -5,7 +5,8 @@ test_that("reverse_geocoding works as expected", {
   reverse_geocoding_result1 <- reverse_geocoding(0.1, 51.5)
   reverse_geocoding_result2 <- reverse_geocoding(0.127, 51.507)
   reverse_geocoding_result3 <- reverse_geocoding("0.1275", "51.5073", limit = 3)
-  reverse_geocoding_result4 <- reverse_geocoding("0.1275", "51.5073", limit = 11, radius = 200)
+  reverse_geocoding_result4 <- reverse_geocoding("0.1275", "51.5073",
+                                                 limit = 11, radius = 200)
 
   expect_that(reverse_geocoding_result2, is_a("list"))
 
