+++ tests/testthat/test-place_query.R
@@ -3,7 +3,7 @@ context("place_query")
 test_that("place_query works as expected", {
   incorrect_limit1 <- 101
   incorrect_limit2 <- "11"
-  correct_limit = 11
+  correct_limit <- 11
 
   place_query_result1 <- place_query("Hills")
   place_query_result2 <- place_query("Hills", limit = correct_limit)
@@ -13,4 +13,5 @@ test_that("place_query works as expected", {
 
   expect_error(place_query("EC1", limit = incorrect_limit1))
   expect_error(place_query("EC1", limit = incorrect_limit2))
+  expect_error(place_query("H"))
 })
