### 2018-01-10 #176 @erzk
- Tests added
- If I use Imports for httr, then most functions crash. This is the main dependency so using Depends is stronger than Imports
- '=' assignments fixed
- long lines were mostly split, except the functions definitions where it's more readable (IMHO) to keep them on one line
- vapply() won't do the job in the highlighted functions
- I imported the httr to avoid calling several httr function in most of PostcodesioR functions
- I don't get the LaTeX error. Not sure what's causing it.
- folder removed
- Additionally, I fixed the DESCRIPTION file by adding the URL to postcodes.io in Description
