### 2018-01-21 #176 @cvitolo
@karthik Here is my review.

@erzk Nice package and works like a charm, well done!
Below are few comments, the vast majority are related to the documentation, only few to functionalities. I hope my comments are self-explanatory but if you need further explaination please do not hesitate to ask :)

---

#### **1. Duplicated column names**
It is good practice to output a table in which column names are unique. If there are duplicates, these columns cannot be easily subsetted by name. A table with duplicated column names is obtained when running the following command (taken from vignette):
lookup_result <- postcode_lookup("EC1Y8LX")
```
In `lookup_result`, columns `admin_district`, `admin_county`, `admin_ward`, `parish`, `parliamentary_constituency`, `ccg` are duplicated. It is not simple to spot those duplicated labels, however, one way to go about this is to convert the data.frame to a tibble:
library(tibble)
lookup_result_tbl <- as.tibble(lookup_result)
```
This way, if you have duplicated column names, you get a self-explanatory error.
In this case it returns:

Error: Columns `admin_district`, `admin_county`, `admin_ward`, `parish`, `parliamentary_constituency`, `ccg`, `nuts` must have unique names.

I would suggest to make all the column names unique.

#### **2. Function Documentation:**
All the exported functions are documented but it would be extremely useful to have more info on the returned values.

#### **3. bulk_postcode_lookup only works with lists with 2 or more elements**
Not sure why `bulk_postcode_lookup` only works with lists with 2 or more elements. Why is that? This limitation should be documented in the help page or, alternatively, you could merge `postcode_lookup` and `bulk_postcode_lookup` in one unique function (this would me my preferred option). The function could check whether the input is a list or a string and then apply `bulk_postcode_lookup` or `postcode_lookup`, respectively.

#### **4. The documentation of bulk_reverse_geocoding needs updating**
The help page of bulk_reverse_geocoding needs updating:
* the title says 'Reverse geocoding' rather then 'Bulk reverse geocoding'
* there should be instructions on how to construct the input for the `geolocations` argument.
* the 'details' section mention that the method requires a JSON object but the `geolocations` argument is a list. You probably meant the API method not this package?

#### **5. Logic of nearest_* functions**
The documentation of `nearest_postcode` should clarify how the nearest postcodes are selected. Is it sorting them based on the distances between postcode's centres?
Same question applies to `nearest_outcode` and `nearest_outcode_lonlat`.

#### **6. Function postcode_autocomplete**
There is a typo in the help page: 'an list' should be corrected to 'a list'

#### **7. Functions not mentioned in vignette**
These functions are not mentioned in the vignette:
* outward_code_lookup
* postcode_autocomplete (there is a typo in the help page: 'an list' should be corrected to 'a list')
* postcode_query
* postcode_validation
* terminated_postcode
I understand these are minor functions but I think at least `postcode_autocomplete` could be mentioned as it is a very useful.

#### **8. postcode_query**
What is the difference between `postcode_query` and `postcode_lookup`?
The documentation should mention what data are retrieved in both cases.

#### **9. terminated_postcode**
Description section needs updating, as it seems to describe one of the `nearest_*` functions.

It should also explain what it is meant by 'terminated'.

#### **10. Query limitation**
Many functions have a query limit of 100, can this lead to misleading results? Please expand and document this limitation in the appropriate help pages.

#### **11. LICENSE**
This is just a general question/curiosity: the Postcodesio API has an MIT license, have you thought of applying the same license rather than a GPL3?

#### **12. Note in check**
When I run devtools::check() on the package I get 1 note:

Status: 1 NOTE
checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘README.Rmd’

This note can be removed by adding README.Rmd to your .Rbuildignore file.

#### **13. DESCRIPTION file**
I noticed you added httr as a dependency. Please consider using the Import field and import only the functions you need, specifying them in the NAMESPACE file, see example below:
importFrom(httr, GET)
```

#### **14. Lines longer than 80 characters**
When I run lintr checks, it shows that the following lines are longer than 80 characters:

* outcode_reverse_geocoding line 22
* reverse_geocoding line 22
* test-outcode_reverse_geocoding lines 14-15-16

Please consider to split each of the above lines into two.

#### **15. Avoid sapply()**
When I run lintr checks, it shows that you use sapply().
If you can, please consider using vapply() instead, as it returns a consistent data type.

#### **16. Vignette**
The vignette runs fast and smoothly, well done!

The content of the vignette is pretty much the same as the README. I would suggest not to duplicate the content because it can be difficult to keep them synched.

If you want to keep a vignette, in my opinion, this should be a bit more descriptive than a README file (they often resemble the structure of journal papers). You could maybe expand the introduction to include information about postcodeio (open source, based exclusively on open data, etc.), the target audience of your package and its potential applications. In my opinion, your package has great potential not only as a stand-alone tool, but also as a back-end tool for shiny applications to map and filter geospatial information. It could also be a useful dependency for other packages, to filter data based on postcodes rather than counties/local authorities/coordinates (I have a couple of them, rnrfa and rdefra, that could make good use of PostcodeioR).
Of course, it would be great to also have a more detailed description of each function (in which case you would use it, how the result compares to what you get from another function and so forth).

#### **17. Function documentation**
All the exported functions are documented but it would be extremely useful to have:

* a sentence to explain that spaces in postcodes are ignored (e.g. for postcode_lookup and bulk_postcode_lookup),
* the returned values for all the functions should be described in greater detail.
