### 2018-01-21 #176 @erzk
@cvitolo Thanks for the comments. Very useful. I'll address the issues that you mentioned ASAP.
