### 2018-02-13 #176 @Robinlovelace
Package Review
--------------

-   \[x\] As the reviewer I confirm that there are no conflicts of
    interest for me to review this work (If you are unsure whether you
    are in conflict, please speak to your editor *before* starting your
    review).

#### Documentation

-   \[ \] **A statement of need** clearly stating problems the software
    is designed to solve and its target audience in README
    -   There are plenty of examples in the README but little on the
        target audience and more contextual problems it's aiming to
        solve ("e.g. you are researching movement patterns based on
        postcode data and want to visualise it") so a little work needed
        here I think. A problem with the README: there is too much code
        output. It's overwhelming.
-   \[ \] **Installation instructions:** for the development version of
    package and any non-standard dependencies in README

    if (!require("devtools")) install.packages("devtools")

    ## Loading required package: devtools

    devtools::install_github("erzk/PostcodesioR")

    ##   Use `force = TRUE` to force installation

    library(PostcodesioR)

    ## Loading required package: httr

works fine but the style is non-standard (expect a PR incoming)

-   \[ \] **Vignette(s)** demonstrating major functionality that runs
    successfully locally
-   The `Introduction.Rmd` vignette seems to duplicate `README.Rmd`. I
    suggest reducing this duplication to a minimum and making the README
    shorter.
-   \[x\] **Function Documentation:** for all exported functions in R
    help
-   Seem well-documented
-   \[x\] **Examples** for all exported functions in R Help that run
    successfully locally
-   \[ \] **Community guidelines** including contribution guidelines in
    the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports`
    and `Maintainer` (which may be autogenerated via `Authors@R`).

r1. UK postcodes are a widely used form of geo-referenced and are seldom
easy to use, so the idea behind this package is welcome. I think it the
approach of wrapping an online service could be useful for small,
intermittent use cases. However for larger datasets, a package that does
not rely on internet access would be more useful. I think this would be
possible, based on this statment in the README:

> [postcodes.io](https://postcodes.io/) provides full list of
> geolocation data that can be used locally without limitations.

r2. Would creating a package that did this be an option?

r3. Regarding style, the package does not adhere to Ropensci's style
guide in a number of ways. I suggest renaming the package `postcodesio`
- shorter and easier to type.

r4. Lack of support for common spatial classes for input and output.
Such as simple feature points.

r5. The output of calls to postcodes.io seems to be overly verbose. I am
confident a data frame output type is more appropriate in this case,
particularly as the result for a single lookup is a data frame:

    lookup_result <- postcode_lookup("EC1Y8LX")
    class(lookup_result)

    ## [1] "data.frame"

    pc_list <- list(postcodes = c("PR3 0SG", "M45 6GN", "EX165BL"))
    bulk_lookup_result <- bulk_postcode_lookup(pc_list)

    class(bulk_lookup_result)

    # str(bulk_lookup_result[1]) # not shown: overly verbose!

r6. At times the suggested input formats are not user friendly. I do not
think this is a realistic user entry, for example:

    # create a JSON object with the coordinates
    geolocations_list <- structure(
     list(
     geolocations = structure(
     list(
     longitude = c(-3.15807731271522, -1.12935802905177),
     latitude = c(51.4799900627036, 50.7186356978817),
     limit = c(NA, 100L),
     radius = c(NA, 500L)),
     .Names = c("longitude", "latitude", "limit", "radius"),
     class = "data.frame",
     row.names = 1:2)),
     .Names = "geolocations")

    bulk_rev_geo <- bulk_reverse_geocoding(geolocations_list)

It would be more useful if the user were given an easy function to
create such inputs into `bulk_reverse_geocoding`

r7. **httr** would be better as an Imports than as a Depends (I'm
finally transitioning away from Depending on **sp** for my **stplanr**
package)

Overall great package, thanks for the opportunity to review it and I
hope this feedback is useful!
