### 2018-02-20 #176 @erzk
@Robinlovelace thanks for the review. I was away for a while so I couldn't work on the package but now I'll get back to it.
