### 2018-03-03 92fded1
description added. examples moved to vignette
