+++ README.md
@@ -3,12 +3,14 @@ PostcodesioR
 
 [![Travis-CI Build Status](https://travis-ci.org/erzk/PostcodesioR.svg?branch=master)](https://travis-ci.org/erzk/PostcodesioR) [![Coverage Status](https://img.shields.io/codecov/c/github/erzk/PostcodesioR/master.svg)](https://codecov.io/github/erzk/PostcodesioR?branch=master) [![Package-License](http://img.shields.io/badge/license-GPL--3-brightgreen.svg?style=flat)](http://www.gnu.org/licenses/gpl-3.0.html) [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/PostcodesioR)](https://cran.r-project.org/package=PostcodesioR)
 
-An API wrapper around [postcodes.io](https://postcodes.io/) - free UK postcode lookup and geocoder.
+An API wrapper around [postcodes.io](https://postcodes.io/) - free UK postcode lookup and geocoder. This package helps to find and transform information about UK administrative geography like postcodes, LSOA, MSOA, constituencies, counties, wards, districts, CCG or NUTS.
+
+The package is based exclusively on open data provided by postcodes.io. PostcodesioR can be used by data scientists or social scientists working with geocoded UK data. A common task when working with such data is aggregating geocoded data on different administrative levels, e.g. turning postcode-level data into counties. This package can help is achieving this and in many other cases.
 
 Installation
 ------------
 
-This package is still in a development version and is available only on GitHub. Not sure if it will ever make it to CRAN. In order to install it use the following commands:
+This package is a developmental version and is available only on GitHub. Hopefully, it will eventually make it to CRAN. In order to install PostcodesioR use the following commands:
 
 ``` r
 if(!require("devtools")) {
@@ -29,13 +31,11 @@ library(PostcodesioR)
 Examples
 --------
 
-Where possible, I tried to return a data frame. Unfortunately, a lot of API calls return more complex data and in those cases it is safer to use lists.
+Where possible, I tried to return a data frame. Unfortunately, a lot of API calls return more complex data and in those cases it is safer to use lists. The API uses defaults to limit the number of returned calls. Check the function documentation for more details.
 
-### Lookup postcodes and outcodes
+For additional information about the returned data and the function calls see the original [documentation](https://postcodes.io/docs).
 
-#### Single postcode
-
-Provide a postcode to obtain all available information:
+The main function of this package provides information related to a given postcode
 
 ``` r
 lookup_result <- postcode_lookup("EC1Y8LX")
@@ -75,754 +75,11 @@ str(lookup_result)
     ##  $ ccg                       : chr "E38000088"
     ##  $ nuts                      : chr "UKI43"
 
-#### Multiple postcodes
-
-``` r
-pc_list <- list(postcodes = c("PR3 0SG", "M45 6GN", "EX165BL"))
-bulk_lookup_result <- bulk_postcode_lookup(pc_list)
-
-#overview
-str(bulk_lookup_result[1])
-```
-
-    ## List of 1
-    ##  $ :List of 2
[…]
