+++ README.Rmd
@@ -8,12 +8,14 @@ output: rmarkdown::github_document
 [![Package-License](http://img.shields.io/badge/license-GPL--3-brightgreen.svg?style=flat)](http://www.gnu.org/licenses/gpl-3.0.html)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/PostcodesioR)](https://cran.r-project.org/package=PostcodesioR)
 
-An API wrapper around [postcodes.io](https://postcodes.io/) - free UK postcode lookup and geocoder.
+An API wrapper around [postcodes.io](https://postcodes.io/) - free UK postcode lookup and geocoder. This package helps to find and transform information about UK administrative geography like postcodes, LSOA, MSOA, constituencies, counties, wards, districts, CCG or NUTS.
+
+The package is based exclusively on open data provided by postcodes.io. PostcodesioR can be used by data scientists or social scientists working with geocoded UK data. A common task when working with such data is aggregating geocoded data on different administrative levels, e.g. turning postcode-level data into counties. This package can help is achieving this and in many other cases.
 
 ## Installation
 
-This package is still in a development version and is available only on GitHub. Not sure if it will ever make it to CRAN.
-In order to install it use the  following commands:
+This package is a developmental version and is available only on GitHub. Hopefully, it will eventually make it to CRAN.
+In order to install PostcodesioR use the following commands:
 
 ```{r, eval=FALSE}
 if(!require("devtools")) {
@@ -32,13 +34,11 @@ library(PostcodesioR)
 
 ## Examples
 
-Where possible, I tried to return a data frame. Unfortunately, a lot of API calls return more complex data and in those cases it is safer to use lists.
+Where possible, I tried to return a data frame. Unfortunately, a lot of API calls return more complex data and in those cases it is safer to use lists. The API uses defaults to limit the number of returned calls. Check the function documentation for more details.
 
-### Lookup postcodes and outcodes
+For additional information about the returned data and the function calls see the original [documentation](https://postcodes.io/docs).
 
-#### Single postcode
-
-Provide a postcode to obtain all available information: 
+The main function of this package provides information related to a given postcode
 
 ```{r, message = FALSE, warning = FALSE}
 lookup_result <- postcode_lookup("EC1Y8LX")
@@ -47,203 +47,11 @@ lookup_result <- postcode_lookup("EC1Y8LX")
 str(lookup_result)
 ```
 
-#### Multiple postcodes
-
-```{r}
-pc_list <- list(postcodes = c("PR3 0SG", "M45 6GN", "EX165BL"))
-bulk_lookup_result <- bulk_postcode_lookup(pc_list)
-
-#overview
-str(bulk_lookup_result[1])
-```
-
-#### Outward code lookup
-
-Provide an outcode to obtain geolocation data for the centroid of the specifid outcode:
-
-```{r}
-ocl <- outward_code_lookup("E1")
-
-#overview
-str(ocl)
-```
-
-### Reverse geocoding
-
-Provide latitude and longitude to obtain geographic information. Different levels of aggregation are available, i.e. postcode or outcode.
-
-#### Single postcode
-
-```{r}
-rev_geo <- reverse_geocoding(0.127, 51.507)
-
-# overview
-str(rev_geo[1])
-```
-
-#### Multiple postcodes
-
-To reverse geocode multiple values use the function underneath. The result is a nested list, which might be a bit intimidating, but it allows storing unequal lengths of elements.
-
-```{r}
-# create a JSON object with the coordinates
[…]
