+++ R/outcode_reverse_geocoding.R
@@ -19,7 +19,8 @@
 #' outcode_reverse_geocoding("-3.15807731271522", "51.4799900627036")
 #' outcode_reverse_geocoding(-3.15, 51.47, limit = 11, radius = 20000)
 #'
-outcode_reverse_geocoding <- function(longitude, latitude, limit = 10, radius = 5000) {
+outcode_reverse_geocoding <- function(longitude, latitude,
+                                      limit = 10, radius = 5000) {
   if (nchar(longitude) < 4) {
     stop("Please provide a valid longitude with at least two decimal points.")
   }
