+++ R/terminated_postcode.R
@@ -1,18 +1,18 @@
 #' Terminated postcode lookup
 #'
-#' Returns nearest postcodes for a given longitude and latitude.
+#' Returns month and year if a postcode was terminated or is no longer active.
 #'
 #' @import httr
 #'
 #' @param postcode A string. Terminated UK postcode.
 #'
-#' @return A data frame with data about terminated postcode.
+#' @return A data frame with data about terminated postcode. NA if active.
 #'
 #' @export
 #'
 #' @examples
 #' terminated_postcode("EC1Y 8LX") # exisitng postcode - returns 404
-#' terminated_postcode("E1W 1UU") # terminated - from Wikipedia but returns 404
+#' terminated_postcode("E1W 1UU") # terminated
 #'
 terminated_postcode <- function(postcode) {
   if (!is.character(postcode) || nchar(postcode) < 3) {
