+++ man/postcode_lookup.Rd
@@ -17,6 +17,6 @@ Lookup a postcode.
 }
 \examples{
 postcode_lookup("EC1Y8LX")
-postcode_lookup("EC1Y 8LX")
+postcode_lookup("EC1Y 8LX") # spaces are ignored
 
 }
