+++ man/bulk_postcode_lookup.Rd
@@ -16,7 +16,8 @@ A list.
 Returns a list of matching postcodes and respective available data.
 }
 \examples{
-pc_list <- list(postcodes = c("PR3 0SG", "M45 6GN", "EX165BL"))
+pc_list <- list(
+postcodes = c("PR3 0SG", "M45 6GN", "EX165BL")) # spaces are ignored
 bulk_postcode_lookup(pc_list)
 
 }
