+++ tests/testthat/test-postcode_autocomplete.R
@@ -14,8 +14,8 @@ test_that("postcode_autocomplete works as expected", {
   expect_error(postcode_autocomplete("E1", limit = incorrect_limit1))
   expect_error(postcode_autocomplete("E1", limit = incorrect_limit2))
 
-  expect_that(postcode_autocomplete_result1, is_a("list"))
-  expect_that(postcode_autocomplete_result2, is_a("list"))
+  expect_that(postcode_autocomplete_result1, is_a("data.frame"))
+  expect_that(postcode_autocomplete_result2, is_a("data.frame"))
 
-  expect_equal(length(postcode_autocomplete_result2), correct_limit)
+  expect_equal(nrow(postcode_autocomplete_result2), correct_limit)
 })
