### 2018-03-03 #176 @erzk
@cvitolo I made changes as you recommended. I answered your comments in more detail:

1. I'm not sure how to address this point. The problem here is that in some cases the API returns a structure that can't be turned into a data frame. I need to work on this a bit longer. TODO: df vs. list
2. Additional information about returned data types was added to README. I think it might be too long to describe each data point in each function as they are repeating. Do you have other suggestion?
3. The functions are split into two different calls by postcodes.io. I tried to keep the function as similar to the function calls as possible. The API specifies that bulk works with two or more postcodes. I added this info to documentation.
4. Title fixed. TODO: JSON vs. list
5. This info is missing from the API documentation. I will contact the authors for more detail.
6. Fixed. Well spotted.
7. Fixed
8. Both function seem to be the same. `postcode_lookup` returns a data frame and `postcode_query` a list. I don't think there was any particular reason for that. I will check with the API authors. `postcode_query` call allows limiting the results. The rest seems the same.
9. Fixed. Documentation is up to date now.
10. Limits of 100 are the default in the API calls so I kept it the same in the package.
11. GPL3 keeps the modifications free (and open source) so I prefer it to the MIT license.
12. Fixed. Thanks for the tip. I had the same problem in my other packages.
13. I haven't changed that as I rely strongly on httr. A helper function in zzz.R is calling three functions from httr and there are additional ones in functions. That's my only dependency so I'm loading it as is. What would be a benefit of importing each function from httr separately?
14. Fixed
15. I used `goodpractice` package which suggested these changes but using `vapply()` breaks my code and would need extra steps to reshape the output of `sapply()`
16. The README files has new information about use cases and explains where the package can be useful. README was trimmed down because @Robinlovelace also suggested it's too long. The examples (except one) were moved to a vignette.
17. Comments added to function documentation (e.g. `postcode_lookup()`). Returned values are described in the original documentation. I refer users to it. The description itself is quite long and I don't think it needs to be repeated. Also, see point 2.

Thanks again for reviewing the code. I still need to address Robin's comments.
