+++ R/zzz.R
@@ -1,3 +1,8 @@
+#' @importFrom httr GET
+#' @importFrom httr user_agent
+#' @importFrom httr warn_for_status
+#' @importFrom httr status_code
+#' @importFrom httr content
 ua <- httr::user_agent("http://github.com/erzk/PostcodesioR")
 
 extract_results <- function(x) {
