### 2018-03-04 #176 @erzk
@Robinlovelace Thanks for the comments, they were very helpful.

**Statement of need** was added.
**README** was shortened and examples moved to a vignette.
**Installation**: what's non-standard about the style? what would you suggest?

r1. I agree.
r2. The next step will be to create a package (or new functions in this package) that can achieve the same thing locally. Any suggestions would be very helpful.
r3. postcodesio would not be search engine friendly. postcodesior would be better, without the capital cases I used but that would mean renaming not only the package but also some blog posts or links. If it's not a major issue, I'd prefer to stick to the old name. I will use the lowercase for future packages.
r4. Can you elaborate? I'm using the basic cases that are allowed/returned by the API.
r5. I agree. Data frames would be easier to use. The problem here was that in some cases the lists that are returned had unequal number of elements thus turning them into a data frame was tricky. Will try to find an example. It was more tricky than expected. PR would be welcome if you have a solution.
r6. I clarified this example and showed how to turned a data frame into a format required by the function.
r7. fixed. @cvitolo suggested the same thing. I've tried it before but I must have done something wrong.
