### 2018-03-18 #176 @cvitolo
@erzk thanks for addressing all my comments. Great job!

Please find below few more suggestions.

1. I'm not sure how to address this point. The problem here is that in some cases the API returns a structure that can't be turned into a data frame. I need to work on this a bit longer. TODO: df vs. list

Maybe your data.frame needs list-columns?

2. Additional information about returned data types was added to README. I think it might be too long to describe each data point in each function as they are repeating. Do you have other suggestion?

I know it's a tedious task but a full description of the expected output is needed. If you have multiple functions (e.g. A, B, C) generating the same table, you could add the full description to function A. Function B and C can simply contain a link to the description in function A. Does it make sense?

Alternatively you can add links to the original API documentation. But please try to add them systematically to every help page.

3. The functions are split into two different calls by postcodes.io. I tried to keep the function as similar to the function calls as possible. The API specifies that bulk works with two or more postcodes. I added this info to documentation.

OK, it makes sense to keep the R functions as similar to the function calls as possible. Do you mind to add a note on this to the package help pages (postcode_lookup and bulk_postcoe_lookup)?

One more thing, at the moment if your `postcode` input argument for postcode_lookup() is a vector of two postcodes, the function fails:

```r
> postcode_lookup(c("EC1Y8LX", "RG13DQ"))
Error: length(url) == 1 is not TRUE
```

It would be great to make this error more meaningful. Maybe you could add something like this at the beginnig of postcode_lookup():

```r
if (length(nchar(postcode)) > 1){
    stop("This function accepts only one postcode, for multiple requests please use bulk_postcode_lookup().")
  }
```

This way, when you run the function with the wrong input, the user gets a clear suggestion to fix the problem:

```r
> postcode_lookup(c("EC1Y8LX", "RG13DQ"))
Error in postcode_lookup(c("EC1Y8LX", "RG13DQ")) :
  This function accepts one postcode, for multiple postcodes please use bulk_postcode_lookup().
```

4. Title fixed. TODO: JSON vs. list

OK, let me know when you fix the JSON vs list bit.

5. This info is missing from the API documentation. I will contact the authors for more detail.

Perfect, let me know when the info is added to the documentation.

6. Fixed. Well spotted.

There is still an inconsistency to fix in the documentation: the Value section says the function returns a list, while in reality it returns a data.frame (as also stated in the Description section).

7. Fixed

OK

8. Both function seem to be the same. postcode_lookup returns a data frame and postcode_query a list. I don't think there was any particular reason for that. I will check with the API authors. postcode_query call allows limiting the results. The rest seems the same.

I would suggest to add a comment on this in the relevant help pages.

9. Fixed. Documentation is up to date now.

OK

10. Limits of 100 are the default in the API calls so I kept it the same in the package.

Is there a way to change the default API limit of 100 calls? If so, is there a way to change that default parameter in your package as well?

11. GPL3 keeps the modifications free (and open source) so I prefer it to the MIT license.

OK

12. Fixed. Thanks for the tip. I had the same problem in my other packages.

No problem :)

13. I haven't changed that as I rely strongly on httr. A helper function in zzz.R is calling three functions from httr and there are additional ones in functions. That's my only dependency so I'm loading it as is. What would be a benefit of importing each function from httr separately?

In my opinion, it is better to 'consciously' import the functions you need because that eliminates the problem of masking functions by mistake.

14. Fixed

OK

15. I used goodpractice package which suggested these changes but using vapply() breaks my code and would need extra steps to reshape the output of sapply()

OK

16. The README files has new information about use cases and explains where the package can be useful. README was trimmed down because @Robinlovelace also suggested it's too long. The examples (except one) were moved to a vignette.

OK

17. Comments added to function documentation (e.g. postcode_lookup()). Returned values are described in the original documentation. I refer users to it. The description itself is quite long and I don't think it needs to be repeated. Also, see point 2.

That's fine. It would be very convenient to have links to the original documentation in the help pages as well.
