+++ DESCRIPTION
@@ -2,7 +2,8 @@ Package: PostcodesioR
 Type: Package
 Title: API wrapper around postcodes.io (free UK postcode lookup and geocoder)
 Version: 0.1.1
-Authors@R: person("Eryk", "Walczak", email = "eryk.j.walczak@gmail.com", role = c("aut", "cre"))
+Authors@R: person("Eryk", "Walczak", email = "eryk.j.walczak@gmail.com", role = c("aut", "cre"),
+    comment = c(ORCID = "0000-0001-8724-462X"))
 Maintainer: Eryk Walczak <eryk.j.walczak@gmail.com>
 Description: Free UK geocoding using data from Office for National Statistics.
     It is using several functions to get information about post codes, outward
