### 2018-03-18 4565b9a
throw an error when more than 1 postcode
