+++ R/postcode_lookup.R
@@ -6,6 +6,7 @@
 #' @importFrom httr warn_for_status
 #'
 #' @param postcode A string. One valid UK postcode.
+#' For more than one postcode use \link[PostcodesioR]{bulk_postcode_lookup}.
 #'
 #' @export
 #'
