+++ man/bulk_postcode_lookup.Rd
@@ -7,7 +7,8 @@
 bulk_postcode_lookup(postcodes)
 }
 \arguments{
-\item{postcodes}{Accepts a list of postcodes. Accepts up to 100 postcodes.}
+\item{postcodes}{Accepts a list of postcodes. Accepts up to 100 postcodes.
+For only one postcode use \link[PostcodesioR]{postcode_lookup}.}
 }
 \value{
 A list.
