### 2018-03-20 #176 @Robinlovelace
Comments post-review:

> r1. I agree. + r2. The next step will be to create a package (or new functions in this package) that can achieve the same thing locally. Any suggestions would be very helpful.

No suggestions at present, other than to download all the boundaries used by the API and use a spatial package such as **sf** to extract the boundaries for any point. This could link to my prototype https://github.com/Robinlovelace/ukboundaries/ package which I hope (with some encouragement from @karthik and perhaps support from the ONS) to submit to ROpenSci at some point.

> r3. postcodesio would not be search engine friendly. postcodesior would be better, without the capital cases I used but that would mean renaming not only the package but also some blog posts or links. If it's not a major issue, I'd prefer to stick to the old name. I will use the lowercase for future packages.

It's not a major issue.

> r4. Can you elaborate? I'm using the basic cases that are allowed/returned by the API.

I think this *is* a major issue. R has some add-on packages that make it intuitive to process, analyse and model spatial data. Premier among these for vector data is **sf**. If it could take `sf` objects in and perhaps return `sf` objects out that would be a great benefit. I suggest taking a look at https://github.com/ropensci/osmdata (and my own **stplanr** package). These allow inputs and outputs in various spatial formats which is very useful if you work with spatial data which presumably users of **PostcodesioR** will.

> r5. I agree. Data frames would be easier to use. The problem here was that in some cases the lists that are returned had unequal number of elements thus turning them into a data frame was tricky. Will try to find an example. It was more tricky than expected. PR would be welcome if you have a solution.

Like @cvitolo I think list columns or empty columns could save you here. Unlikely to find time to put in a PR to do this with the priority being to complete the https://github.com/Robinlovelace/geocompr book at present.

> r6. I clarified this example and showed how to turned a data frame into a format required by the function.

Great work.

> r7. fixed. @cvitolo suggested the same thing. I've tried it before but I must have done something wrong.

Awesome.
