+++ R/postcode_lookup.R
@@ -6,11 +6,94 @@
 #' @importFrom httr warn_for_status
 #'
 #' @param postcode A string. One valid UK postcode. This function is case- and space-insensitive.
-#' For more than one postcode use \link[PostcodesioR]{bulk_postcode_lookup}.
+#' For more than one postcode use \code{\link{bulk_postcode_lookup}}.
 #'
 #' @export
 #'
 #' @return A data frame. Returns all available data if found. Returns 404 if postcode does not exist.
+#'  * \code{postcode} Postcode. All current ('live') postcodes within the United Kingdom,
+#'  the Channel Islands and the Isle of Man, received monthly from Royal Mail.
+#'  2, 3 or 4-character outward code, single space and 3-character inward code.
+#'  * \code{quality} Positional Quality. Shows the status of the assigned grid reference.
+#'    * 1 - within the building of the matched address closest to the postcode mean
+#'    * 2 - as for status value 1, except by visual inspection of Landline maps (Scotland only)
+#'    * 3 - approximate to within 50m
+#'    * 4 - postcode unit mean (mean of matched addresses with the same postcode, but not snapped to a building)
+#'    * 5 - imputed by ONS, by reference to surrounding postcode grid references
+#'    * 6 - postcode sector mean, (mainly PO Boxes)
+#'    * 8 - postcode terminated prior to Gridlink(R) initiative, last known ONS postcode grid reference1
+#'    * 9 - no grid reference available
+#'  * \code{eastings} Eastings. The Ordnance Survey postcode grid reference
+#'  Easting to 1 metre resolution; blank for postcodes in the
+#'  Channel Islands and the Isle of Man.
+#'  Grid references for postcodes in Northern Ireland relate to the Irish Grid system.
+#'  * \code{northings} Northings. The Ordnance Survey postcode grid reference
+#'  Easting to 1 metre resolution; blank for postcodes in the
+#'  Channel Islands and the Isle of Man.
+#'  Grid references for postcodes in Northern Ireland relate to the Irish Grid system.
+#'  * \code{country} Country. The country (i.e. one of the four constituent
+#'  countries of the United Kingdom or the Channel Islands or the Isle of Man)
+#'  to which each postcode is assigned.
+#'  * \code{nhs_ha} Strategic Health Authority. The health area code for the postcode.
+#'  * \code{longitude} Longitude.
+#'  The WGS84 longitude given the Postcode's national grid reference.
+#'  * \code{latitude} Latitude.
+#'  The WGS84 latitude given the Postcode's national grid reference.
+#'  * \code{european_electoral_region} European Electoral Region (EER).
+#'  The European Electoral Region code for each postcode.
+#'  * \code{primary_care_trust} Primary Care Trust (PCT).
+#'  The code for the Primary Care areas in England, LHBs in Wales, CHPs in Scotland,
+#'  LCG in Northern Ireland and PHD in the Isle of Man;
+#'  there are no equivalent areas in the Channel Islands.
+#'  Care Trust/ Care Trust Plus (CT) / local health board (LHB) /
+#'  community health partnership (CHP) / local commissioning group (LCG) /
+#'  primary healthcare directorate (PHD).
+#'  * \code{region} Region (formerly GOR). The Region code for each postcode.
+#'  The nine GORs were abolished on 1 April 2011 and are now known as 'Regions'.
+#'  They were the primary statistical subdivisions of England and
+#'  also the areas in which the Government Offices for the Regions fulfilled their role.
+#'  Each GOR covered a number of local authorities.
+#'  * \code{lsoa} 2011 Census lower layer super output area (LSOA).
+#'  The 2011 Census lower layer SOA code for England and Wales,
[…]
