### 2018-04-07 #176 @erzk
@cvitolo
1. Fixed. The column names are not duplicated any more.
2. Fixed. I added full documentation and URLs to `postcode_lookup()`, `place_lookup()`, and `terminated_postcode()`. They return three different sets of values as described in the API [documentation](#' @seealso \code{\link{place_lookup}} for documentation.). The remaining functions have references to one of these three relevant functions.
3. Fixed. Extra documentation added. Error handling added.
4. Fixed. Documentation update. R uses a list. API uses JSON.
5. Fixed. The author updated the API documentation. So did I. Nearest postcodes/outcodes are based on outcode/postcode centroids.
6. Fixed. Documentation says 'data frame'.
8. Fixed. Documentation updated to explain differences between the functions.
10. Fixed. Where possible `@param limit` was added to specify the limit.
13. Fixed. `@importFrom httr GET` was added.
17. Fixed. See point 2.
