### 2018-04-16 #176 @karthik
Thanks for the revisions @erzk and for the detailed and thoughtful reviews @cvitolo and @Robinlovelace
Can you two look over the latest revisions and sign off if you are satisfied? 🙏
