### 2018-04-16 #176 @karthik
>  If I think of any further issues would they be best as issues in the issue tracker or as comments here?

It would be best on the issue tracker of the repo itself (the location will change once accepted, but GitHub will redirect appropriately).
