### 2018-04-17 #176 @cvitolo
I'm satisfied with the changes, well done @erzk!
