### 2018-04-17 #176 @karthik
Congrats @erzk! your package has been approved! 🎉 Thank you for submitting and @cvitolo and @Robinlovelace  for thorough and timely reviews. To-dos:

```
[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)

```
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.

Final suggestion. If you found feedback from the reviewers valuable, consider adding them as reviewers in the description of your package (this is entirely up to you and not a requirement).  If you decide to, you'll need:

- The ORCID for both reviewers (if the reviewers don't have it, they should sign up for one here: https://orcid.org/)
- Follow this format: https://github.com/ropensci/rdflib/blob/master/DESCRIPTION#L8-L10
- More context for why such attribution matters: https://github.com/ropensci/rdflib/blob/master/DESCRIPTION#L8-L10
