### 2018-04-17 #176 @erzk
@karthik All done.
@stefaniebutland  I'm happy to write a blog post about the package.
@cvitolo @Robinlovelace Thanks for reviewing the code. I added you as reviewers in the DESCRIPTION file.
