### 2018-04-18 #176 @stefaniebutland
@erzk Great to hear you're interested in contributing a post about `PostcodesioR`. You can read example narrative posts about onboarded packages here: https://ropensci.org/tags/onboarding/ and technotes here: https://ropensci.org/technotes/. Technotes can be published at any time, while blog posts are scheduled in advance.

Here are some technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post.

We ask that you submit a draft via pull request at least one week before the planned publication date. At this point there are several blog posts in the pipeline, so perhaps best for you to suggest a deadline by which you would like to submit a draft.
